将命名错误改正
PointFragment -----> ProjectFragment
PointList           -----> PointListFragment
