+++ app/src/main/java/com/gzcz/rtchen/positionmarker/PointListFragment.java
@@ -13,12 +13,12 @@ import android.widget.TextView;
 /**
  * A simple {@link Fragment} subclass.
  * Activities that contain this fragment must implement the
- * {@link PointList.OnFragmentInteractionListener} interface
+ * {@link PointListFragment.OnFragmentInteractionListener} interface
  * to handle interaction events.
- * Use the {@link PointList#newInstance} factory method to
+ * Use the {@link PointListFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class PointList extends Fragment {
+public class PointListFragment extends Fragment {
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
     private static final String ARG_PARAM1 = "param1";
@@ -33,7 +33,7 @@ public class PointList extends Fragment {
     View mView = null;
     TextView list_view_text_view = null;
 
-    public PointList() {
+    public PointListFragment() {
         // Required empty public constructor
     }
 
@@ -43,11 +43,11 @@ public class PointList extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment PointList.
+     * @return A new instance of fragment PointListFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static PointList newInstance(String param1, String param2) {
-        PointList fragment = new PointList();
+    public static PointListFragment newInstance(String param1, String param2) {
+        PointListFragment fragment = new PointListFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -68,7 +68,7 @@ public class PointList extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
 
-        mView = inflater.inflate(R.layout.point_fragment, container, false);
+        mView = inflater.inflate(R.layout.pointlist_fragment, container, false);
         list_view_text_view = (TextView) mView.findViewById(R.id.list_view_text_view);
 
 //        String prj = getActivity().getIntent().getStringExtra("prj");
