--- app/src/main/res/layout/point_fragment.xml
@@ -1,34 +1,0 @@
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context="com.gzcz.rtchen.positionmarker.PointFragment">
-
-    <!-- TODO: Update blank fragment layout -->
-    <LinearLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:id="@+id/linearLayout">
-
-        <EditText
-            android:id="@+id/edit_view_prj"
-            android:layout_width="300dp"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"/>
-
-        <Button
-            android:id="@+id/btn_add_prj"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:textAllCaps="false"
-            android:text="Add"/>
-    </LinearLayout>
-
-    <ListView
-        android:layout_width="wrap_content"
-        android:layout_height="464dp"
-        android:id="@+id/list_view_prj"
-        android:layout_gravity="center_horizontal|bottom" />
-
-</FrameLayout>
