+++ app/src/main/res/layout/pointlist_fragment.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.gzcz.rtchen.positionmarker.PointList">
+    tools:context="com.gzcz.rtchen.positionmarker.PointListFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <LinearLayout
