+++ app/src/main/res/layout/project_fragment.xml
@@ -1,10 +1,34 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="vertical" android:layout_width="match_parent"
-    android:layout_height="match_parent">
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context="com.gzcz.rtchen.positionmarker.ProjectFragment">
 
-    <TextView
+    <!-- TODO: Update blank fragment layout -->
+    <LinearLayout
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="This is Project Fragment."/>
-</LinearLayout>
+        android:id="@+id/linearLayout">
+
+        <EditText
+            android:id="@+id/edit_view_prj"
+            android:layout_width="300dp"
+            android:layout_height="wrap_content"
+            android:layout_weight="1"/>
+
+        <Button
+            android:id="@+id/btn_add_prj"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_weight="1"
+            android:textAllCaps="false"
+            android:text="Add"/>
+    </LinearLayout>
+
+    <ListView
+        android:layout_width="wrap_content"
+        android:layout_height="464dp"
+        android:id="@+id/list_view_prj"
+        android:layout_gravity="center_horizontal|bottom" />
+
+</FrameLayout>
