+++ app/src/main/res/menu/activity_main_drawer.xml
@@ -12,9 +12,9 @@
             android:icon="@drawable/ic_menu_gallery"
             android:title="Gallery" />
         <item
-            android:id="@+id/nav_point"
+            android:id="@+id/nav_project"
             android:icon="@drawable/ic_menu_slideshow"
-            android:title="Point" />
+            android:title="Project" />
         <item
             android:id="@+id/nav_manage"
             android:icon="@drawable/ic_menu_manage"
